Conflicts:
	tita-business/src/main/java/at/ac/tuwien/ifs/tita/dao/interfaces/IEffortDao.java
	tita-business/src/main/java/at/ac/tuwien/ifs/tita/dao/time/EffortDao.java
	tita-business/src/sql/ins_base_data.sql
	tita-business/src/test/java/at/ac/tuwien/ifs/tita/dao/test/time/EffortDaoTest.java
	tita-dao/src/main/java/at/ac/tuwien/ifs/tita/business/service/project/IProjectService.java
	tita-dao/src/main/java/at/ac/tuwien/ifs/tita/business/service/project/ProjectService.java
	tita-entity/src/main/java/at/ac/tuwien/ifs/tita/entity/IssueTrackerProject.java

+++ tita-dao/src/main/java/at/ac/tuwien/ifs/tita/business/service/time/IEffortService.java
@@ -84,7 +84,7 @@ public interface IEffortService {
      * 
      * @return list of years
      */
-    List<Integer> getEffortsYears();
+    List<Double> getEffortsYears();
 
     /**
      * Gets a view for the last time efforts.
